Move to standalone component to fix type issues
