+++ frontend/src/toolbar/experiments/WebExperimentTransformHeader.tsx
@@ -7,6 +7,8 @@ import { SelectorCount } from '~/toolbar/actions/SelectorCount'
 import { experimentsTabLogic } from '~/toolbar/experiments/experimentsTabLogic'
 import { WebExperimentTransform } from '~/toolbar/types'
 
+import { SelectorEditor } from './SelectorEditor'
+
 interface WebExperimentTransformHeaderProps {
     variant: string
     transformIndex: number
@@ -34,7 +36,10 @@ export function WebExperimentTransformHeader({
             </div>
 
             {transform.selector && (
-                <SelectorCount selector={transform.selector} variant={variant} transformIndex={transformIndex} />
+                <div className="float-right flex items-center gap-2">
+                    <SelectorCount selector={transform.selector} />
+                    <SelectorEditor selector={transform.selector} variant={variant} transformIndex={transformIndex} />
+                </div>
             )}
 
             {experimentForm?.variants && (
